mentre/requirements.txt
fastparquet==2024.5.0
numpy==1.26.4
pandas==2.1.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==1.4.53
sqlalchemy-redshift==0.8.14